bug deletedestination id_to undefined

+++ delete.php
@@ -2,6 +2,13 @@
 session_start();
 include './classes/TourOperator.php';
 $newTO = new TourOperator();
-$newTO->delete();
 
+if(isset ($_POST['submitTO'])){
+$newTO->delete();
 header('Location: ./admin.php?id_admin'.$_SESSION['id_admin']);
+}
+
+if(isset($_POST['submitDest'])){
+$newTO->deleteDestination();
+header('Location: ./toPage.php?id_to'.$_SESSION['id_to']);
+}
